refactor AMG creation as a class
