+++ opm/simulators/linalg/bda/opencl/openclCPR.hpp
@@ -28,6 +28,7 @@
 #include <opm/simulators/linalg/bda/opencl/opencl.hpp>
 #include <opm/simulators/linalg/bda/opencl/openclBILU0.hpp>
 #include <opm/simulators/linalg/bda/Matrix.hpp>
+#include <opm/simulators/linalg/bda/CprCreation.hpp>
 #include <opm/simulators/linalg/bda/opencl/OpenclMatrix.hpp>
 #include <opm/simulators/linalg/bda/opencl/openclPreconditioner.hpp>
 
@@ -39,7 +40,7 @@ template<class Scalar> class BlockedMatrix;
 
 /// This class implements a Constrained Pressure Residual (CPR) preconditioner
 template<class Scalar, unsigned int block_size>
-class openclCPR : public openclPreconditioner<Scalar,block_size>
+class openclCPR : public openclPreconditioner<Scalar,block_size>, public CprCreation<Scalar, block_size>
 {
     using Base = openclPreconditioner<Scalar,block_size>;
 
@@ -54,13 +55,8 @@ class openclCPR : public openclPreconditioner<Scalar,block_size>
     using Base::err;
 
 private:
-    int num_levels;
-    std::vector<Scalar> weights, coarse_vals, coarse_x, coarse_y;
-    std::vector<Matrix<Scalar>> Amatrices, Rmatrices; // scalar matrices that represent the AMG hierarchy
     std::vector<OpenclMatrix<Scalar>> d_Amatrices, d_Rmatrices; // scalar matrices that represent the AMG hierarchy
-    std::vector<std::vector<int> > PcolIndices; // prolongation does not need a full matrix, only store colIndices
     std::vector<cl::Buffer> d_PcolIndices;
-    std::vector<std::vector<Scalar>> invDiags; // inverse of diagonal of Amatrices
     std::vector<cl::Buffer> d_invDiags;
     std::vector<cl::Buffer> d_t, d_f, d_u; // intermediate vectors used during amg cycle
     std::unique_ptr<cl::Buffer> d_rs;      // use before extracting the pressure
@@ -70,33 +66,9 @@ private:
     std::once_flag opencl_buffers_allocated;  // only allocate OpenCL Buffers once
 
     std::unique_ptr<openclBILU0<Scalar,block_size>> bilu0;                    // Blocked ILU0 preconditioner
-    BlockedMatrix<Scalar>* mat = nullptr;    // input matrix, blocked
-
-    using DuneMat = Dune::BCRSMatrix<Dune::FieldMatrix<Scalar, 1, 1> >;
-    using DuneVec = Dune::BlockVector<Dune::FieldVector<Scalar, 1> >;
-    using MatrixOperator = Dune::MatrixAdapter<DuneMat, DuneVec, DuneVec>;
-    using DuneAmg = Dune::Amg::MatrixHierarchy<MatrixOperator, Dune::Amg::SequentialInformation>;
-    std::unique_ptr<DuneAmg> dune_amg;
-    std::unique_ptr<DuneMat> dune_coarse;       // extracted pressure matrix, finest level in AMG hierarchy
-    std::shared_ptr<MatrixOperator> dune_op;    // operator, input to Dune AMG
-    std::vector<int> level_sizes;               // size of each level in the AMG hierarchy
-    std::vector<std::vector<int> > diagIndices; // index of diagonal value for each level
-    Dune::UMFPack<DuneMat> umfpack;             // dune/istl/umfpack object used to solve the coarsest level of AMG
-    bool always_recalculate_aggregates = false; // OPM always reuses the aggregates by default
-    bool recalculate_aggregates = true;         // only rerecalculate if true
-    const int pressure_idx = 1;                 // hardcoded to mimic OPM
-    unsigned num_pre_smooth_steps;              // number of Jacobi smooth steps before restriction
-    unsigned num_post_smooth_steps;             // number of Jacobi smooth steps after prolongation
 
     std::unique_ptr<openclSolverBackend<Scalar,1>> coarse_solver; // coarse solver is scalar
     bool opencl_ilu_parallel;                   // whether ILU0 operation should be parallelized
-
-    // Analyze the AMG hierarchy build by Dune
-    void analyzeHierarchy();
-
-    // Analyze the aggregateMaps from the AMG hierarchy
-    // These can be reused, so only use when recalculate_aggregates is true
-    void analyzeAggregateMaps();
 
     // Initialize and allocate matrices and vectors
     void init_opencl_buffers();
@@ -108,8 +80,6 @@ private:
     void apply_amg(const cl::Buffer& y, cl::Buffer& x);
 
     void amg_cycle_gpu(const int level, cl::Buffer &y, cl::Buffer &x);
-
-    void create_preconditioner_amg(BlockedMatrix<Scalar>* mat);
 
 public:
     openclCPR(bool opencl_ilu_parallel, int verbosity);
@@ -131,12 +101,6 @@ public:
     bool create_preconditioner(BlockedMatrix<Scalar>* mat,
                                BlockedMatrix<Scalar>* jacMat) override;
 };
-
-// solve A^T * x = b
-// A should represent a 3x3 matrix
-// x and b are vectors with 3 elements
-template<class Scalar>
-void solve_transposed_3x3(const Scalar* A, const Scalar* b, Scalar* x);
 
 } // namespace Opm::Accelerator
